a playbook to add an OSD to an existing cluster
The playbook uses the ceph-fetch-keys role to connect to the mons and
retrieve keys for the OSD bootstrapping. Ensure that all your mons for
the cluster you're adding the OSD to is in your inventory in the mons
group. This ensures that a proper ceph.conf is created for the new OSD.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>